Ignore unknown properties when deserializing JSON from clients
We will use the defined/generated POJOs to describe what properties we do/do not
care about when fetching data from pinhead. This avoids the constant errors we are
getting when the pinhead API changes.
